Added basic logging, comments, and fixed SQL statement
I wrote a bad SQL query, so it should be fixed now. It needed to
return the cards matching a search whether they were in the deck
or not. It was doing that, but also returning cards that were in
other decks, and two results if the card was in your editing deck.

The logging is basic, it prints the query to the console for the
Professor's easy viewing of what is being executed.

The comments are just to break up the end points so it's easier
to follow.