Athyng/front end (#8)
* Added /api/:table/:attribute/:value endpoint

* Added card search to front end

* Removed .idea folder from repo

* Updated readme's, added buttons for card removal/addition

* Created decks view, added ability to save a deck

* Added CRUD for cards and decks

* Added basic logging, comments, and fixed SQL statement

I wrote a bad SQL query, so it should be fixed now. It needed to
return the cards matching a search whether they were in the deck
or not. It was doing that, but also returning cards that were in
other decks, and two results if the card was in your editing deck.

The logging is basic, it prints the query to the console for the
Professor's easy viewing of what is being executed.

The comments are just to break up the end points so it's easier
to follow.

* Fixed cards sorting, reinstated endpoint

Whoops. It turns out that end point was used, it's back now. The
cards were being sorted by quantity in ascending order, which
is weird. Changed it to descending followed by alphabetical
name sorting.